Implemented login verification email message, isolated sendEmail method for posible future use.
