+++ src/hooks/emails.ts
@@ -1,0 +1,28 @@
+'use server';
+import nodemailer from 'nodemailer';
+
+// Initialize nodemailer transporter
+const transporter = nodemailer.createTransport({
+  service: 'gmail',
+  auth: {
+    user: process.env.EMAIL_USER,
+    pass: process.env.EMAIL_PASS,
+  },
+});
+
+
+export async function sendEmail(email: string, subject: string, body: string) {
+  try {
+    await transporter.sendMail({
+      from: process.env.EMAIL_USER,
+      to: email,
+      subject: subject,
+      text: body,
+    });
+
+    return { success: true, message: 'Email sent' };
+  } catch (error) {
+    console.error('Error sending email:', error);
+    return { success: false, message: 'Error sending email' };
+  }
+}
